requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
et_xmlfile==2.0.0
folium==0.18.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.13.2
numpy==2.1.3
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
protobuf==5.28.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.10.0
pyproj==3.7.0
python-calamine==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxlsb==1.0.10
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
streamlit_folium==0.23.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
xlrd==2.0.1
XlsxWriter==3.2.0
xyzservices==2024.9.0
